+++ psdi_data_conversion/file_io.py
@@ -7,10 +7,28 @@ Functions and classes related to general filesystem input/output
 
 import glob
 import os
+from functools import lru_cache
 from shutil import copyfile, make_archive, unpack_archive
 from tempfile import TemporaryDirectory
 
 from psdi_data_conversion import constants as const
+
+
+@lru_cache(maxsize=1)
+def get_package_path() -> str:
+    """Gets the absolute path to where the `psdi_data_conversion` package is on disk
+
+    Returns
+    -------
+    str
+    """
+
+    # For an interactive shell, __file__ won't be defined for this module, so use the constants module instead
+    reference_file = os.path.realpath(const.__file__)
+
+    package_path = os.path.dirname(reference_file)
+
+    return package_path
 
 
 def is_archive(filename: str) -> bool: